Fix overlapping item IDs
In the game data items and cargos are stored separately, and sometimes their IDs overlap. In recipes there's a special field that defines whether ID belongs to an item or a cargo.
For BitPlanner, another approach was chosen. All game IDs are in range of UInt32, so BitPlanner will use Uint64 with maximum value of UInt32 being an offset for cargo IDs.

Closes #6